requirements.txt
streamlit>=1.32.0
moviepy>=2.0.0
pillow>=10.0.0
edge-tts>=6.1.10
faster-whisper>=1.0.3
python-dotenv>=1.0.1
requests>=2.31.0
